Attempt to address various issues that the C++ compiler on the build machine warns about (mostly unused variables).

+++ Test/c++/ExternDefs.h
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-typedef unsigned long long uint64;
+typedef uint64_t uint64;
 
 namespace Extern {
   struct state {
